refactor: move low-hanging transforms to own file + docs
